[FIX] accountant_knowledge: fix crash when previewing account report
When previewing an article that contains an account report, the preview
currently crashes. This happens because the preview generator encounters
an unknown `accountReport` block. To render properly, this block must be
transformed - specifically, the reference in `data-embedded-props` must
be resolved and the Python dictionary must be converted into a JSON object.

This commit updates the system to correctly handle `accountReport` blocks
during both preview and article generation. It will also refactor the code
to reduce code redundancy and support the `articleIndex` block across all
templates.

Task-4989809

closes odoo/enterprise#93531

Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>