[FIX] ai_model_field_selector_popover: fixed field picker duplicates
before this commit:

When you press Enter in the field picker, it will always pick the same field even if it is not visible.

Long field paths couldn't be removed because of text moving out of the box size.

After this commit:

Field can only be selected once and it stays visible in the list.
Added a visual check to show that the field is already selected.

Made text wrap in the box properly so long field paths could be removed.

Task-4918680

closes odoo/enterprise#95672

X-original-commit: 0762e363c42522955b92d254f82d6f44f3b96930
Signed-off-by: Adrien Schoffeniels (adsc) <adsc@odoo.com>
Signed-off-by: Basit Arshad (baar) <baar@odoo.com>