[IMP] account_reports: group returns by type on the dashboard
Late returns and returns which deadline is reached in the following month are shown on the return card on the dashboard to remind users to review and submit them.
There was no returns limit. In extreme cases, the return card can be totally overflown - see Spain.

We now group returns by types on the dashboard ; for each type, we display the oldest deadline. The rest of the behaviour remains unchanged.

task-5066144

closes odoo/enterprise#93934

Signed-off-by: Hugo Poncelet (hupo) <hupo@odoo.com>