[IMP] stock_barcode: unpack button
Add an "Unpack" button on package lines in the Barcode app.
When this button is clicked, we remove the outermost result package, or
free the result package move lines from their package dest.
The unpack barcode (OBTUPCK) is also added to the `barcodes_actions.pdf`
file.

task-5065193

Part-of: odoo/enterprise#94091
Signed-off-by: Quentin Wolfs (quwo) <quwo@odoo.com>