[FIX] appointment: remove copy from capacity inverse
[1] fixes an issue with not being able to copy calendar event
with their booking lines by also adding the capacity as inverse.

If it was ever necessary it is not anymore and we can get rid of
nonsensical context keys guarding the inverse by removing it.
Leaving the test introduced with the fix untouched.

[1]: 6e85fff3d9fce23f8591e6bb3340749c889c9934

Part-of: odoo/enterprise#92263
Related: odoo/odoo#222741
Related: odoo/upgrade#8262
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>