[FIX] iot: prevent duplicated actions on fallback
In the following case an IoT box action can be duplicated:
1. Request is sent over longpolling
2. Action takes longer than 6s to execute, longpolling times out on the
   client side
3. Websocket request is sent as a fallback
4. Both requests cause an action to be executed (double blackbox
   registration, double print, etc.)

To solve this we will send an ID with every action, and if the ID has
already been recently seen we will ignore the action and log a warning.

This PR handles the client side code, which just involves adding the
new `action_unique_id` key to the `iot_http` fallbacks.

task-5067737

closes odoo/enterprise#94398

X-original-commit: ebdf27918096f4d42496674b8244b77a1faf3575
Related: odoo/odoo#226490
Signed-off-by: Max Whale (mawh) <mawh@odoo.com>