[FIX] account_reports: exchange rate entries in bank reconciliation
report

**Steps to reproduce**
- Have foreign currency with rates for date 1 and date 2
- Have a Bank journal in foreign currency
- Register a transaction in date 1
- In date 2 open the unrealized currency report
- Create the adjustment entry
- From the Accounting dashboard Bank[EUR] > Reconciliation report

**Issue**
The adjustment entry difference is present under the 'Misc.
operations' line.
This occurs because we look for journal entries hitting the bank account
but that specific entry should not be reported as it does not represent
a bank in/out operation

A solution is to exclude the exchange entry journal, so any operation
reported there is not taken into account in the report

opw-4867870

closes odoo/enterprise#94263

X-original-commit: fa4243dec0610053ab63204c47b02ceb208633a6
Signed-off-by: Hugo Poncelet (hupo) <hupo@odoo.com>
Signed-off-by: Andrea Grazioso (agr) <agr@odoo.com>