[IMP] crm_enterprise, helpdesk: hide measure `Stage Color`
This commit hides the measure `Stage Color` from graph, cohort and pivot
views of the `crm.lead, helpdesk.ticket` model, as it doesn't really make
sense to measure on a stage color.

The measures are made from the model's fields and all possible measures
are shown unless it is made invisible. Therefore make it explicitly
invisible.

Task-5065924

closes odoo/enterprise#93992

Related: odoo/odoo#225687
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>