[FIX] hr_payroll: traceback due to a warning message
Problem
----------
The warning message compares the contract_date_start of the version with the date_from of the payslip.
So if no contract is defined on the versions, it raises the traceback

Solution
----------
Remove this warning message it is useless we only want to know if a contract exists for the version associated with the payslip.
Fix the "Contract" Button to open the employee form view with the good version.

task-5067543

closes odoo/enterprise#94006

Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>