[FIX] sign: load fonts before rendering font selection
When clicking on "auto" in the signature modal the state of the modal was
updated before the fonts were loaded. In some cases this leads to the modal
rerendering in "auto" mode without the fonts having loaded yet. As the fonts
are not part of the state this leads to all of the fonts (except the default)
being missing.

Part-of: odoo/enterprise#88518
Related: odoo/odoo#215789
Related: odoo/upgrade#8207
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>