[IMP] appointment: add resource filters on apt calendar view
Purpose
=======
Removing the attendee filter on the resource based appointment calendar view
as filtering on resource based events using the attendees is tedious.
Add a filter on the linked appointment type resources instead.

Specification
=============
The attendee filter isn't adequate when talking about resource based bookings.
The bookings aren't displayed if you don't set one of its attendee in the filters
and looking through attendees using the "+ Add attendee" widget in the sidebar
is not a viable solution.

Technical
=========
Removing all the 'filters', 'write_model', 'write_field', 'filter_field'
and 'avatar_field' attributes from the 'partner_ids' field as using
any of those automatically add the field in the filters
(they are OR-ed to add the filter info in the calendar arch parser).

Task-5059001

closes odoo/enterprise#94037

Related: odoo/odoo#225785
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>