[FIX] ai,ai_crm: unsudo model instances and env in tool calls
`ai.agent` object is accessed via sudo in the `/ai/generate_response` endpoint.
The sudo is necessary to gain access to the `ai.agent` and related models like
`ai.topic` and tools. However, when tool calls are performed, we want to ensure
that the execution is in user's context.

Intentionally, lead creation via the ai tool is allowed for public and portal
users, thus, explicit sudo in the ai tool's code.

closes odoo/enterprise#92870

Signed-off-by: Panagiotis Kyriakou (paky) <paky@odoo.com>