[IMP] documents: improve the responsiveness of the cogwheel
The cogwheel which holds the folder actions was slow to open because it loads
the actions at startup. To speedup it up, we backport the fix
odoo/enterprise#90124 that allows the cogwheel to be opened while loading the
actions (instead of waiting that the actions are loaded). We also add a spinner
while it is being loaded.

When selecting action to embed for the folder, it was slow as well. To solve
the problem, we toggle the action immediately (not waiting the answer of the
server) and roll it back in case of failure.

Finally, to limit the number of calls to the server, we only reload the search
model after a delay if the user has not toggled another action in the meantime
(using the debounce method).

Task-4828503

closes odoo/enterprise#96545

X-original-commit: 8e7036443d0e9ad29e9b5f6c6f1e2743465b196b
Signed-off-by: Stéphane Debauche (std) <std@odoo.com>
Signed-off-by: Pierre-Yves Dufays (pydu) <pydu@odoo.com>