[IMP] hr_payroll: make kanban payrun buttons consistent with payslip list view
This commit changes the visibility conditions of buttons in the payrun
kanban view to be consistent with the payslip list view filtered by
payruns. Moreover, since the `PayslipListController` is using
`PayRunCard` component, which uses `StatusBubble` component that
requires information of the payrun model fields, `payslips_with_issues`
and `has_error` fields are added to the kanban view to be extracted in
`PayslipListController` component and injected to `PayRunCard` as props,
which fixes the inconsistent status color.

task-5067711

closes odoo/enterprise#94024

Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>