[FIX] account_budget: only show budget of current companies
Steps to reproduce:
- Install Accounting
- In Accounting settings, activate "Budget Management"

- Go to "Accounting / Accounting / Analytic Budget"
- Create a budget
- Open the budget
- Create a bill (or invoice depending on the budget type) using
the analytic used in the budget lines

- Create a budget from another company and also a bill using the
analytic used in the budget lines

- Connect with a user with Accounting rights who only has access
to the first company

- Go to "Accounting / Reporting / Management / Budget Report"
- Switch to the list view
- Remove the "Open Budget" filter

Issue:
The entries linked to the budget from the other company are
visible.
When trying to open one of them, an access error is raised.

Solution:
Add a multi-company record rule for budget report.

opw-4943360

closes odoo/enterprise#94927

X-original-commit: fdc5694517c0f2411df25bd61d5315468f56399b
Signed-off-by: William André (wan) <wan@odoo.com>
Signed-off-by: Anh Thao Pham (pta) <pta@odoo.com>