[IMP] helpdesk: add rotting fields
Helpdesk tickets and ticket stages are modified to take advantage
of the new rotting logic of the mail.tracking.duration mixin, allowing
to get a better sense of when a task gets stale.

When a ticket spends too long without seeing a stage or kanban
state change, it will appear in a different color in kanban views,
to highlight it and bring it to attention.

task-5064952

closes odoo/enterprise#93901

Related: odoo/odoo#225502
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>