[IMP] data_cleaning: re-enable and convert to hoot unit test
PR [1] merged the data_merge addon, which defined a qunit test
and properly declared it in the correct bundle in its manifest,
which the data_cleaning addond. Although the test file has been
kept, it was no longer declared in the manifest of the new addon,
so the test was never run anymore.

This commit re-enables it and converts it to hoot.

[1] https://github.com/odoo/enterprise/pull/62979

closes odoo/enterprise#94049

Signed-off-by: Aaron Bohy (aab) <aab@odoo.com>