[FIX] account_reports: search_bar filtering on no data
Recently, we changed reports to load without their data (and lines),
since the search bar is loaded with the options it will try to filter
on lines that haven't been loaded.

Now, we are not searching if the lines are not loaded and search again
once they are loaded.

task-5068340

closes odoo/enterprise#94057

Signed-off-by: Hugo Poncelet (hupo) <hupo@odoo.com>