[IMP] ai: add random prompts in systray
Purpose:
--------
When opening the ai chat using the systray button, one wants to show 3
different examples of prompts buttons, based on the current modules that
are installed.

In order to not create a lot of bridge modules that would each contain a
a few lines of data, these data are installed with the ai module and the
search method of the ai prompt button model has been overridden to skip
records from uninstalled modules.

Task-5008938

Part-of: odoo/enterprise#94059
Signed-off-by: Joseph Caburnay (jcb) <jcb@odoo.com>