[IMP] website: Introduce high-level cache for website pages
This commit introduces a high-level cache on the `website.page` model,
similar to the one that existed before the `t-cache` directive was
implemented in `ir.qweb`. This cache is designed to significantly improve
performance by storing the HTML response of published pages, thereby
reducing rendering time and database queries.

The management and logic for this cache are placed directly on the
`website.page` model, as it is the source of the HTML response for these
records. This approach makes it easy to control caching behavior through
a few new, overridable methods:

-  `_allow_to_use_cache`: Determines whether a page is allowed to be
served from the cache based on the current request, URL, or session. For
example, a page with a `sale.order` in the session would not use the cache.

-  `_allow_cache_insertion`: Checks if the generated HTML content is
eligible for caching. This is useful for preventing sensitive or dynamic
content from being stored. For instance, a page with a reference to a
`sale.order` would not be inserted into the cache.

-  `_post_process_response_from_cache`: A hook called after a response is
retrieved from the cache. This method allows for post-processing, such as
incrementing counters or modifying HTTP headers, without regenerating the
entire page.

see https://github.com/odoo/odoo/pull/224487
see https://github.com/odoo/odoo/pull/88276

closes odoo/enterprise#94487

Related: odoo/odoo#226627
Related: odoo/design-themes#1151
Signed-off-by: Raphael Collet <rco@odoo.com>