[FIX] whatsapp: prevent fails in command palette tests
Before this commit, clicking on some elements of the command palette
would not trigger the associated event, this was probably caused
by clicking on outdated owl fragments which cannot trigger events while
being destroyed. This commit fixes this issue by making sure that
the interface is in a stable configuration before clicking.

https://runbot.odoo.com/odoo/runbot.build.error/112148

closes odoo/enterprise#94397

X-original-commit: a757167ea343789879b0973d44f6e0be59b8215f
Related: odoo/odoo#226472
Signed-off-by: Sébastien Theys (seb) <seb@odoo.com>
Signed-off-by: Thanh Son Dodeur (tso) <tso@odoo.com>