[FIX] web_gantt: disable drop for rows grouped by readonly field
Issue
-----
Gantt view's drag & drop allows the user to change the value of readonly fields
if they are stored.

E.G. in the Planning view of MRP, grouped by Work Center > Product, dragging &
dropping can change the product of the WO if the user is not careful and drops
the WO on top of another product's WO.

Steps to reproduce
-----
- Have 2 products
- Create a MO for product 1 with a WO at work center 1, plan it
- Create a MO for product 2 with a WO at work center 2, plan it
- Got to Manufacturing, Planning, Planning by Work Center
- Add a custom group (by product)
- Drag the WO of WC2 and drop it on top of the other WO

> Both the WC and the product of the second WO change

Cause
-----
The example problem is only for versions 17.0 & 18.0 where the `product_id`
field of `mrp.workorder` is both readonly and stored.

https://github.com/odoo/odoo/blob/31e46a841b38de0f99beb1844f985bc670621486/addons/mrp/models/mrp_workorder.py#L34

While the user cannot change the field value manually, automatic actions such as
a gantt view drag & drop can change its' value by passing it to `write` since
the field is stored. This does not pose any problem for related fields that are
not stored.

More broadly, gantt views should not ignore the `readonly` attribute of fields.

Solution
-----
Add a new `o_gantt_readonly` class to all cells of rows grouped by a readonly
field - and their "child" rows. For example, if the grouping is done by
"Work Center > Product > Quality Check" and "Product" is readonly, rows grouped
by either "Product" or "Quality Check" will be marked as readonly.

When the user drags a pill, dynamically remove the class from cells of the same
"child group". The class will then be added back upon pill drop.

-----
Ticket:
opw-4875366

closes odoo/enterprise#100209

X-original-commit: 33c97a746c28f363cbb4435a6a81ce957addcea1
Signed-off-by: David Van Droogenbroeck (drod) <drod@odoo.com>