[IMP] ai: add instruction to have times in user's timezone
Purpose:
--------
Currently, only the current time is given to the LLM using the user's tz.
When asking for example for the next event in the calendar, the answer
will be using UTC.

Instead, we now give the current time in UTC (as the records data will be
sent in UTC), and instruct the model to use the user's tz when generating
responses.

Task-5008938

closes odoo/enterprise#94059

Signed-off-by: Joseph Caburnay (jcb) <jcb@odoo.com>