[FIX] l10n_in_reports: prevent traceback on tds/tcs excel opening
With this commit:
https://github.com/odoo/odoo/commit/f6d6659f8bf9ab7c7409f584524c7509865a7a45
The `l10n_in_pan` field was removed and replaced with the
`l10n_in_pan_entity_id` field, which relates to the `l10n_in.pan.entity` model.

However, the fix introduced in that commit wasn’t updated accordingly when the
field name changed.

With this commit, the field will now be set correctly, and the traceback will no
longer occurs when generating the TDS/TCS Excel sheet.

closes odoo/enterprise#94156

Signed-off-by: Josse Colpaert (jco) <jco@odoo.com>