[REF] ai_website_livechat: refactor ai_livechat website snippet
1. Remove s_ai_livechat_data div

The config data of the s_ai_livechat website snippet was being stored
twice, one in the dataset of the ai_livechat snippet element and another time
as data attributes on the 's_ai_livechat_data' div. To clean this
redundency, the s_ai_livechat data div was removed along with all the
references to it.

2. Ref fallback button options

Group the options fallbackButtonText and fallbackButtonURL under
fallbackButton category in snippet options.

3. Refactor ai_livechat component

The ai_livechat component has complex template/logic. This commit aims
at refactoring the component to make it more simple.

Part-of: odoo/enterprise#93686
Related: odoo/odoo#225089
Signed-off-by: Panagiotis Kyriakou (paky) <paky@odoo.com>