[FIX] account_accountant: test multiple lines exchange move
When selecting multiple lines in the bank rec widget (reconcile button), it
could happen that one of those lines have a exchange diff move linked to it.
In this case, the exchange move id was placed on the first line all the time
which could be wrong.

This commit will change the use of indexes to use the reconciled line of the
exchange diff move.

no task id

closes odoo/enterprise#95620

X-original-commit: 2ae5530a22108952288ddeb55efaf3dafd5c802f
Related: odoo/odoo#228780
Signed-off-by: Wala Gauthier (gawa) <gawa@odoo.com>
Signed-off-by: Maximilien La Barre (malb) <malb@odoo.com>