[FIX] account_reports: analytic account filter not shown
The origin of the bug comes from this https://github.com/odoo/enterprise/pull/86716

It added a `_init_options_filters` function which was called with a default sequence of 200. The `_init_options_analytic_groupby` has a sequence of 995. The _init_options_filters basically ran before the analytic account options which it depends on.

Th fix was to set a sequence large enough to `_init_options_filters` so that it would always run after everything.

closes odoo/enterprise#94170

Signed-off-by: Hugo Poncelet (hupo) <hupo@odoo.com>