[FIX] l10n_ec_edi: accept invoice's XML with tag `CDATA`
Issue:
Ecuadorian customer can download their invoice's XML from the SRI but the file contains the invoice in a wrapper tag (`<![CDATA[ ... ]]>`) that prevent the extraction of the data to fill the vendor bill form view.

Steps to reproduce:
- Install "l10n_ec_edi" and switch to an Ecuadorian company
- Have a file downloaded from the SRI.
- Go to Accounting > Vendor > Bills
- Click "Upload" and select the file
- The generated move is not populated with the data

Cause:
Odoo is expecting the XML to not be in the tag `CDATA` and it gets ignored.

Solution:
The change is in `_get_import_file_type` to detect the new type of file as `'l10n_ec.factura'`.

The CDATA content can be fetched by getting the content of the tag `comprobante`.
We then try to convert the content of `comprobante` to XML. If it's possible we have an XML on which we can do the same check as before to know if it's an Ecuadorian invoice.
We then replace the `file_data['xml_tree']` by the content of `CDATA` to have the correct XML for the data extraction.

opw-5004636

closes odoo/enterprise#95036

X-original-commit: 599f9ed025273e668c88d4bee3a790b5712f2399
Signed-off-by: Antoine Boonen (aboo) <aboo@odoo.com>
Signed-off-by: Mathieu Coutant (mcou) <mcou@odoo.com>