[FIX] ai: ask ai button should open the ask ai chat window
The issue is caused by the use of orm service wrapped in "useService".
When the button is clicked, it will be unmounted but an orm query is
still pending. The pending query is blocked by the useService because
the component is already unmounted. The remedy is to use the raw orm
service.

closes odoo/enterprise#94417

Signed-off-by: Louis Baudoux (lba) <lba@odoo.com>