[FIX] account_iso20022: Skip tests for accountant only tests
The aim of this commit is skipping 2 newly added tests from
this commit [[1]]. These 2 tests are useful when accountant
module is installed because payments are, per default, created
without entry in this setup. Without this module, payments
are created with an entry and then don't need to be checked.

runbot-error-231650
runbot-error-231651

[1]: https://github.com/odoo/enterprise/commit/7c36a5f405a0426bf11eca0e5a8a6cf6a416076c

closes odoo/enterprise#94192

Signed-off-by: Wala Gauthier (gawa) <gawa@odoo.com>