[FIX] appointment: follow up ui changes in appointment
In recent work [1] and [2] we improved the ui of the appointment app, respectively
adding a complete form view for questions, that are now a m2m field, and buttons
in the kanban view to quickly access the appointment form view (e.g. 'Configure').

In doing so, we missed two issues:

- [1] When creating a question from the appointment form view, the placeholder
field is focused when opening the form. Set default_focus on the name so that
it is focused instead (issue probably related to having a widget on the name)

- [2] The 'Configure' button, when website is installed and the appointment is
published, is not having a correct click area. The issue comes from the 'published'
ribbon widget. It's taking the height of the card, even though not obvious, hence
capturing the click event. The right end of the button is therefore not working and
the click is instead the default one on the card, redirecting to the kanban view.
Instead of moving the button more to the left, as the view is already having a lot
of information and squeezing elements would be difficult, we instead give the button
a 1 z-index.

[1] odoo/enterprise@09be817f0e4ec12fbf0a795cbf9225e3e4746a4e
[2] odoo/enterprise@6199b6baa56a2ca7742835345633af6a00549075

Task-5057504

Part-of: odoo/enterprise#94209
Related: odoo/odoo#226115
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>