[FIX] account_reports: include expense_other in P&L
In saas-18.3 a new account type was added: "Other Expenses"
These accounts are excluded from the account  many2one field to make it easier to find relevant expense accounts for vendor bills.

Issue: This account type is not included in the Profit & Loss Report.

Since l10n_be is the only CoA which uses this account type, the issue was not uncovered until OXP testing days.

closes odoo/enterprise#100074

X-original-commit: 1dea0a6616dbc8872371091fb13c0cc503117256
Signed-off-by: Hugo Poncelet (hupo) <hupo@odoo.com>