[FIX] web_enterprise: avoid flickering on first load in system-dark
As the user's theme now defaults to "system" (aka. the
device/system/browser light/dark preference), when it actually matches
the "dark" theme and the user logs in for the first time, a reload has
to be performed to apply the correct theme.

While this reload is performed quite early in the WebClient boot process
(cf. start of the services), it may actually fully boot before the
reload is applied, presenting a partially usable HomeMenu (light) to the
user. But as the reload actually switches it to its "dark" variant, the
user sees a flickering between the two mode.

This commit limits this effect by completely blocking the WebClient
startup during the reload, avoiding this "light-to-dark flicker" and
giving a better sense of responsiveness.

task-5074720

closes odoo/enterprise#94223

Related: odoo/odoo#226231
Signed-off-by: Adrien Dieudonné (adr) <adr@odoo.com>