[FIX] ai_documents_account: can not save the sort wizard for Finances
Bug
===
When trying to save the sort wizard for a folder having a pinned multi
action, with a `documents_account_record_create` child action, then
the constraint `_check_use_in_ai` was triggered, and it shouldn't.

The reason is the way recursive compute work, we can not set `False`
as the default, because it will allow the parent to be computed with
that.

Task-5077569

Part-of: odoo/enterprise#94290
Signed-off-by: Patrick Hoste (pko) <pko@odoo.com>