[FIX] account_reports: TB when applying a filter when component is destroyed
For the currency rate filter, applyFilters is called when the dropdown is closed.
It might heppen that we have the dropdown open then we click on something that trigger the destruction of the report view but that still trigger applyFilters which try to apply filter on a destroyed component which is illegal.

To prevent this we check if the component is destroyed before reloading the controller.

closes odoo/enterprise#94530

X-original-commit: 88566ba68634da4f641ac218163c263ff452a565
Signed-off-by: Hugo Poncelet (hupo) <hupo@odoo.com>
Signed-off-by: Logan Staelens (lost) <lost@odoo.com>