[FIX] ai_documents: don't add non "use in AI" by default
Purpose
=======
Some action can not be executed by the LLM. For example,
"Link to vehicle" opens a modal to choose the record.

To mitigate this, we don't set the non "use in AI" by default,
but the users can still select them manually if needed.

When the tool tries to execute a window action, we raise an error
so the user can know that the action won't work.

Task-5077569

Part-of: odoo/enterprise#94290
Signed-off-by: Patrick Hoste (pko) <pko@odoo.com>