[IMP] stock_barcode: remove package type
Remove the package type writes aside of the package's name on the
barcode lines, because often, the package types use their own sequence
and this sequence is used when a package of this package type is
created, so the information is most of the time redundant.
Also, now multiple level of packages are display on a line, it's not
clear if the package type is for the package / the result package or for
the parent package / the outermost package.

task-5065193

closes odoo/enterprise#94091

Signed-off-by: Quentin Wolfs (quwo) <quwo@odoo.com>