[IMP] accountant_knowledge: avoid unnecessary page titles in audit reports
Previously, the algorithm that generated the audit report PDF would
create a page title for empty articles using the article name. This
often resulted in unnecessary pages, since many template articles are
considered empty (because they contain elements that should not be printed).

To reduce page clutter, we are revising the algorithm and introducing a
new heuristic: a page title will be generated only if the article
contains a single heading (h1, h2, or h3) and no other text elements.
In such cases, the heading itself will be used as the page title.

Task-5079472

closes odoo/enterprise#94327

Signed-off-by: Damien Abeloos (abd) <abd@odoo.com>