[FIX] appointment: avoid user taking capacity in resource meetings
In [1], we introduced capacity for users. Therefore, capacity is
handled by using booking lines, with capacity reserved / used set
on them. However, unlike the resources, the user is not stored and
is related as the responsible (user_id) of the meeting instead.

This has a side effect: if there are resource booking lines on a
meeting, and a responsible is set (manually or the appointment type
creator as a default responsible, as done in appointment module),
then those booking lines may end up counting as booked capacity for
the responsible as well!

Therefore, we now make sure that the appointment type linked to the
meeting is scheduled based on users when searching for booking lines
in the remaining capacity computation.

STEPS TO REPRODUCE
==================

0. Create a new db and install website_appointment
1. Log in as mitchell Admin
2. Create an appointment_type R based on resources, with a resource R1
and an appointment_type U based on users, with only Mitchell Admin set
as staff user
3. Log out and book R on a given slot S
4. Once done, go to appointment U and try to book the same slot S
5. Once you submit the form, you are brought back to the calendar
selection screen with an error message (code 'failed-staff-user')

TEST
====

A test is added covering most cases, both when getting slots and
when directly measuring the user's remaining capacity.

1: odoo/enterprise@bce7e94650c337a9046958a7689c81db5b2a4c73

Task-5080374

closes odoo/enterprise#94637

X-original-commit: ef50d289a3d8a4c5f00002367abed105145f4555
Signed-off-by: Jérémy Hennecart (jeh) <jeh@odoo.com>
Signed-off-by: Noé Antoine (nan) <nan@odoo.com>