[FIX] documents: fix add document attachment
When adding attachment from documents in the composer, link the attachment to
the composer and not to the thread as it must be linked to the thread only once
the message is posted.

How to reproduce:
- Install the app documents and crm
- Open a lead
- In the chatter click on "Log a note"
- Then click on "Add from Documents"
- Select a document and click on "Add from Documents"
- Reload the page without posting the message
The attachment selected in document is now linked to the lead which shouldn't
be the case.

Note that if you do the same for an expense, as the attachment is linked to the
expense right away when added, the preview panel open immediately, and you have
to reopen "Log a note". That was the original bug detected.

Task-5075835

closes odoo/enterprise#94594

Signed-off-by: Stéphane Debauche (std) <std@odoo.com>