[REF] ai_website_livechat: update `assets_edit_frontend` bundle name
The new html_builder is used by both website and mass_mailing. While the
old name `[html_builder|website].assets_edit_frontend` makes sense in
the context of the website builder, where the iframe is indeed used to
"edit the frontend", it is not the case in mass_mailing, where the email
being composed is not a "frontend". A more generic name is more
appropriate.

This commit is linked to the community PR that updates the bundle name
everywhere.

closes odoo/enterprise#94435

Related: odoo/odoo#226530
Signed-off-by: Francois Georis (fge) <fge@odoo.com>