[FIX] account_{reports,intrastat}: reset button on returns
Two buttons "Reset" are shown in the card of an account return.
This is due to the invisible attributes that wrongly computes the
value. Indeed, it is not possible to write something like
"type_id.is_intrastat_return_type" in an invisible condition, so we
need to make a related field.

closes odoo/enterprise#94393

Signed-off-by: Olivier Colson (oco) <oco@odoo.com>