[IMP] l10n_ph_reports: test fix
Fixes an issue with tests that would fail when the order of
the lines in the file where not consistent.
We will be a bit more precise in the order by of the export
queries, and for SAWT/QAP we will also sort the lines after
grouping per schedules to ensure that the order is consistent
there too.

See runbot error 232581

closes odoo/enterprise#94405

Signed-off-by: William André (wan) <wan@odoo.com>