[FIX] account_accountant: Remove unwanted attachment display on reco
Problem
---------
In odoo/enterprise#85991, attachments from reconciled movs were
displayed in the chatter along side the moves' attachments. This change
was meant to affect only the bank reco widget. However, its scope
reached the standard accounting behavior.

STEPS
---------
1. Have 'account' installed
2. Create a move
3. Confirm and create the PDF attachment from the Send&Print
4. Create a credit note from it
5. Confirm the creadit note
-> The 2 moves are now reconciled and the PDF of the invoice is shown in
the chatter of the credit note

Objective
---------
Only show reconciled moves' attachments in the bank reco chatter.

Solution
---------
Add a context key that is added when the Bank Reco widget is created and
remove it when it is deleted.

opw-#######

closes odoo/enterprise#94661

X-original-commit: 3135e0ba0d78a40e75ffb5d94d6cd6ab1c8a022d
Signed-off-by: Florian Gilbert (flg) <flg@odoo.com>
Signed-off-by: Antoine Boonen (aboo) <aboo@odoo.com>