[FIX] web_enterprise: fix studio blacklist
Steps to reproduce
==================

- Install knowledge, web_studio
- Go to knowledge > Articles
- Click on the optional columns dropdown
- Click on "Add Custom Field"

=> This action is not editable by Studio

Cause of the issue
==================

Since https://github.com/odoo/odoo/pull/205129/commits/c18e93191c516f6e37093f068e8934cba8495891
the event ACTION_MANAGER:UI-UPDATED is triggered before the list
renderer has a chance to register for it.

Solution
========

We don't actually need to listen to that event to get the info we need.
The action xml_id was missing, but it was added in the view subenv
config.

opw-5066488

closes odoo/enterprise#95495

X-original-commit: 23c37b367545d72fe07a3e3302612207a4cec672
Related: odoo/odoo#228472
Signed-off-by: Lucas Perais (lpe) <lpe@odoo.com>
Signed-off-by: Hubert Van De Walle <huvw@odoo.com>