[FIX] l10n_mx_edi_stock: Fix test.
In the PR, the test was changed to comply with the new changes. In that test, a new vehicle was created using a model from the demo data. Since the demo data is not present in all databases, this caused a problem when the ID could not be found. To fix this, we now create a new model using a brand that exists in the data file and will therefore be present in all databases.

See: https://runbot.odoo.com/odoo/runbot.build.error/231693?menu_id=405

I confirm I have signed the CLA and read the PR guidelines at www.odoo.com/submit-pr

closes odoo/enterprise#94495

Signed-off-by: William André (wan) <wan@odoo.com>