[FIX] account_asset: properly display property edit warning
Versions
--------
- saas-18.3

Steps
-----
1. Create a new product without a category;
2. try to edit properties.

Issue
-----
An empty notification appears.

Cause
-----
The `_getPropertyEditWarning` override in `account_asset` introduced by
commit c9db2e6 doesn't return the value of the `super` call.

This was introduced in a forward port as the logic to display the
warning was changed in `web` in saas-18.3, and not requiring a return
value in previous versions.

Solution
--------
Return the `super` call.

opw-4980006

closes odoo/enterprise#94831

X-original-commit: 2ec92d4e424cffa15e75afcb95b2ea83cb37f2ae
Signed-off-by: Paolo Gatti (pgi) <pgi@odoo.com>